W 15125668.separate read write pool (#853)
* W-15125668: Implementation for secondary rocksdb to separate CarbonJ read and write

+++ carbonj.service/src/main/docker/files/entrypoint.sh
@@ -34,8 +34,13 @@ if [ -d /data ]; then
 	mkdir $SERVICEDIR/work/carbonj-data
 	mkdir $SERVICEDIR/work/carbonj-tmp
 	mkdir $SERVICEDIR/work/carbonj-staging
-	mkdir $SERVICEDIR/work/log
-	ln -s $SERVICEDIR/work/log $SERVICEDIR/log
+	if [ "${ROCKSDB_READONLY}" == "false" ]; then
+	  mkdir $SERVICEDIR/work/log
+	  ln -s $SERVICEDIR/work/log $SERVICEDIR/log
+	else
+	  mkdir $SERVICEDIR/work/log-readonly
+    ln -s $SERVICEDIR/work/log-readonly $SERVICEDIR/log
+	fi
 fi
 
 #set default spring profile if none is given. exclusively used for properties overrides file
@@ -53,6 +58,10 @@ for SVC_PROP in `compgen -A variable | grep "^SVC_PROP_"` ; do
 	var_lowercase=${var_space_replaced,,}
 	printf '%s=%s\n' "$var_lowercase" "${!SVC_PROP}" >> $SERVICEDIR/config/overrides.properties
 done
+
+if [ "${ROCKSDB_READONLY}" == "true" ] && [ "${DISABLE_NAMESPACE_COUNTER_CHECK}" == "true" ]; then
+  printf '%s=%s\n' "metrics.store.query.disableNameSpaceCounterCheck" "true" >> $SERVICEDIR/config/overrides.properties
+fi
 
 YOURKIT_PROFILER_OPTS=
 YOURKIT_PROFILER_AGENT_FILE=/app/lib/libyjpagent.so
